requirements.txt
flask
flask-cors
torch
numpy
google-generativeai
sentence-transformers
faiss-cpu
tf-keras
python-dotenv

